https://github.com/morganstanley/ComposeUI/issues/47 created code skeleton
