+++ Tryouts/Prototypes/NoPluginsPrototype/App.axaml
