+++ Tryouts/Prototypes/NoPluginsPrototype/MainWindow.axaml
