+++ Tryouts/Prototypes/NoPluginsPrototype/MainWindow.axaml.cs
