+++ Tryouts/Prototypes/NoPluginsPrototype/Program.cs
