+++ Tryouts/Prototypes/TimeViewPrototype/MainWindow.axaml
