+++ Tryouts/Prototypes/TimeViewPrototype/MainWindow.axaml.cs
@@ -1,0 +1,29 @@
+/// ********************************************************************************************************
+///
+/// Morgan Stanley makes this available to you under the Apache License, Version 2.0 (the "License").
+/// You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0.
+/// See the NOTICE file distributed with this work for additional information regarding copyright ownership.
+/// Unless required by applicable law or agreed to in writing, software distributed under the License
+/// is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+/// See the License for the specific language governing permissions and limitations under the License.
+/// 
+/// ********************************************************************************************************
+
+using Avalonia.Controls;
+using MorganStanley.ComposeUI.Tryouts.Plugins.Services.TimeService;
+using MorganStanley.ComposeUI.Tryouts.Plugins.ViewViewModelPlugins.TimeViewModel;
+
+namespace MorganStanley.ComposeUI.Prototypes.TimeViewPrototype
+{
+    public partial class MainWindow : Window
+    {
+        public MainWindow()
+        {
+            TimeConsumerViewModel timeViewModel = new TimeConsumerViewModel(TimeServicesFactory.CreateTimeService());
+
+            this.DataContext = timeViewModel;
+
+            InitializeComponent();
+        }
+    }
+}
