--- Tryouts/src/Plugins/ViewViewModelPlugins/TimeViewModel/MorganStanley.ComposeUI.Plugins.TimeViewModel.csproj
@@ -1,37 +1,0 @@
-<!--
-********************************************************************************************************
-
-Morgan Stanley makes this available to you under the Apache License, Version 2.0 (the "License").
-You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0.
-See the NOTICE file distributed with this work for additional information regarding copyright ownership.
-Unless required by applicable law or agreed to in writing, software distributed under the License
-is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and limitations under the License.
-
-********************************************************************************************************
--->
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-	
-  <Target Name="PostBuild" AfterTargets="PostBuildEvent">
-  	<Exec Command="xcopy &quot;$(OutDir)\$(AssemblyName).dll&quot; &quot;$(SolutionDir)\bin\$(Configuration)\net6.0\Plugins\ViewModelPlugins\$(ProjectName)\&quot; /R /Y /I" />
-  	<Exec Command="xcopy &quot;$(OutDir)\$(AssemblyName).pdb&quot; &quot;$(SolutionDir)\bin\$(Configuration)\net6.0\Plugins\ViewModelPlugins\$(ProjectName)\&quot; /R /Y /I" />
-  </Target>
-	
-  <ItemGroup>
-    <PackageReference Include="NP.Utilities" Version="1.0.13" />
-    <PackageReference Include="System.Reactive" Version="5.0.0" />
-  </ItemGroup>
-	
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\Core\MorganStanley.ComposeUI.Tryouts.Abstractions\MorganStanley.ComposeUI.Tryouts.Abstractions.csproj" />
-    <ProjectReference Include="..\..\..\Core\MorganStanley.ComposeUI.Tryouts.BasicModels\MorganStanley.ComposeUI.Tryouts.BasicModels.csproj" />
-    <ProjectReference Include="..\..\..\Core\MorganStanley.ComposeUI.Tryouts.Utilities\MorganStanley.ComposeUI.Tryouts.Utilities.csproj" />
-  </ItemGroup>
-
-</Project>
