+++ src/AvaloniaShellApp/App.axaml.cs
