+++ src/AvaloniaShellApp/AvaloniaShellMainWindow.axaml.cs
