+++ src/AvaloniaShellApp/MorganStanley.ComposeUI.AvaloniaShellApp.sln
@@ -1,0 +1,75 @@
+﻿# Morgan Stanley makes this available to you under the Apache License, Version 2.0 (the "License"). You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0. See the NOTICE file distributed with this work for additional information regarding copyright ownership. Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.32112.339
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MorganStanley.ComposeUI.AvaloniaShellApp", "MorganStanley.ComposeUI.AvaloniaShellApp.csproj", "{BCE13BDA-270B-478F-9E11-1CC027DD8147}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{44AEE361-98BE-47C5-94CE-3851C09129A4}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Plugins", "Plugins", "{3224C0CA-18A4-4E10-A5B4-D282DAAC97E5}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MorganStanley.ComposeUI.BasicModels", "..\Core\BasicModels\MorganStanley.ComposeUI.Core.BasicModels.csproj", "{B948F483-BBD0-4E23-A505-E19BCED44297}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MorganStanley.ComposeUI.Utilities", "..\Core\Utilities\MorganStanley.ComposeUI.Core.Utilities.csproj", "{0D8FC7CA-C6EF-4CB9-B165-F8DB8B61E7C1}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MorganStanley.ComposeUI.Avalonia.VisualUtils", "..\Visuals\Avalonia\VisualUtils\MorganStanley.ComposeUI.Visuals.Avalonia.VisualUtils.csproj", "{B7C7C6B2-B896-4781-A19C-6422AFC70944}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{2F3E88D8-966F-4DB1-8E90-3F3DDF0A5515}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Visuals", "Visuals", "{AA0D08E1-FB9B-42AD-A99C-777CF7C2DE6B}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Avalonia", "Avalonia", "{89B16966-AAA4-42F9-B7FF-6C9FB83B2FF3}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "javascript-binding", "javascript-binding", "{978D676B-DDC9-4423-9E9E-422900D17112}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MorganStanley.ComposeUI.Abstractions", "..\Core\Abstractions\MorganStanley.ComposeUI.Core.Abstractions.csproj", "{5A9939BA-2E2B-4A03-9107-AE1F90A4BC93}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Services", "Services", "{E8F066F3-6817-4E45-8FB5-5CB099CC80BD}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{BCE13BDA-270B-478F-9E11-1CC027DD8147}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BCE13BDA-270B-478F-9E11-1CC027DD8147}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BCE13BDA-270B-478F-9E11-1CC027DD8147}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BCE13BDA-270B-478F-9E11-1CC027DD8147}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B948F483-BBD0-4E23-A505-E19BCED44297}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B948F483-BBD0-4E23-A505-E19BCED44297}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B948F483-BBD0-4E23-A505-E19BCED44297}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B948F483-BBD0-4E23-A505-E19BCED44297}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0D8FC7CA-C6EF-4CB9-B165-F8DB8B61E7C1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0D8FC7CA-C6EF-4CB9-B165-F8DB8B61E7C1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0D8FC7CA-C6EF-4CB9-B165-F8DB8B61E7C1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0D8FC7CA-C6EF-4CB9-B165-F8DB8B61E7C1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B7C7C6B2-B896-4781-A19C-6422AFC70944}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B7C7C6B2-B896-4781-A19C-6422AFC70944}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B7C7C6B2-B896-4781-A19C-6422AFC70944}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B7C7C6B2-B896-4781-A19C-6422AFC70944}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5A9939BA-2E2B-4A03-9107-AE1F90A4BC93}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5A9939BA-2E2B-4A03-9107-AE1F90A4BC93}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5A9939BA-2E2B-4A03-9107-AE1F90A4BC93}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5A9939BA-2E2B-4A03-9107-AE1F90A4BC93}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{3224C0CA-18A4-4E10-A5B4-D282DAAC97E5} = {44AEE361-98BE-47C5-94CE-3851C09129A4}
+		{B948F483-BBD0-4E23-A505-E19BCED44297} = {2F3E88D8-966F-4DB1-8E90-3F3DDF0A5515}
+		{0D8FC7CA-C6EF-4CB9-B165-F8DB8B61E7C1} = {2F3E88D8-966F-4DB1-8E90-3F3DDF0A5515}
+		{B7C7C6B2-B896-4781-A19C-6422AFC70944} = {89B16966-AAA4-42F9-B7FF-6C9FB83B2FF3}
+		{2F3E88D8-966F-4DB1-8E90-3F3DDF0A5515} = {44AEE361-98BE-47C5-94CE-3851C09129A4}
+		{AA0D08E1-FB9B-42AD-A99C-777CF7C2DE6B} = {44AEE361-98BE-47C5-94CE-3851C09129A4}
+		{89B16966-AAA4-42F9-B7FF-6C9FB83B2FF3} = {AA0D08E1-FB9B-42AD-A99C-777CF7C2DE6B}
+		{978D676B-DDC9-4423-9E9E-422900D17112} = {44AEE361-98BE-47C5-94CE-3851C09129A4}
+		{5A9939BA-2E2B-4A03-9107-AE1F90A4BC93} = {2F3E88D8-966F-4DB1-8E90-3F3DDF0A5515}
+		{E8F066F3-6817-4E45-8FB5-5CB099CC80BD} = {3224C0CA-18A4-4E10-A5B4-D282DAAC97E5}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {AC408A30-6542-4A25-A6DE-450E325BDE69}
+	EndGlobalSection
+EndGlobal
