+++ src/AvaloniaShellApp/Program.cs
