+++ src/Core/Abstractions/MorganStanley.ComposeUI.Core.Abstractions.csproj
