+++ src/Core/Utilities/MorganStanley.ComposeUI.Core.Utilities.csproj
